+++ src/assets/svg/SvgIndex.jsx
@@ -40,7 +40,7 @@ export const ChevronRight = memo(() => (
     className="chevronRight"
     width="16"
     height="25"
-    viewBox="0 0 16 25"
+    viewBox="0 0 16 30"
     fill="none"
     xmlns="http://www.w3.org/2000/svg"
   >
@@ -54,7 +54,7 @@ export const ChevronLeft = memo(() => (
   <svg
     width="16"
     height="25"
-    viewBox="0 0 16 25"
+    viewBox="0 0 16 30"
     fill="none"
     xmlns="http://www.w3.org/2000/svg"
   >
