+++ src/scss/header/header.scss
@@ -128,63 +128,12 @@
   }
 
   .hamburger{
-    display: flex;
-    position: absolute;
-    top: 1rem;
-    left: 1rem;
-    height: fit-content;
-    flex-direction: column;
-    z-index: 999;
-    transition: 0.2s ease-in-out;
-    cursor: pointer;
-    :nth-child(1){
-      transform: translateY(-0.2rem);
-    }
-    :nth-child(3){
-      transform: translateY(0.2rem);
-    }
-
-    &__line{
-      background-color: $blue-algolia;
-      height: 2px;
-      width: 1rem;
-    }
+    @include hamburger-menu;
   }
-
   .hamburger-active{
-    transition: 0.3s ease-in-out;
-    :nth-child(1){
-      transform: rotateZ(45deg) translateX(1px);
-      transform-origin: center center;
-      transition: 0.3s ease-in-out;
-    }
-    :nth-child(2){
-      transform: rotateZ(-45deg) translateX(1px);
-      transform-origin: center center;
-      transition: 0.3s ease-in-out;
-    }
-    :nth-child(3){
-      opacity: 0;
-      transition: 0.3s ease-in-out;
-    }
+  @include hamburger-menu-active;
   }
   .hamburger-inactive{
-    transition: 0.3s ease-in-out;
-    :nth-child(1){
-      transform: rotateZ(0deg);
-      transform: translateY(-0.2rem);
-      transform-origin: center center;
-      transition: 0.3s ease-in-out;
-    }
-    :nth-child(2){
-      transform: rotateZ(0deg);
-      transform-origin: center center;
-      transition: 0.3s ease-in-out;
-    }
-    :nth-child(3){
-      opacity: 1;
-      transform: translateY(0.2rem);
-      transition: 0.3s ease-in-out;
-    }
+ @include hamburger-menu-inactive;
   }
 }
