+++ src/scss/index.scss
@@ -71,6 +71,6 @@ h3 {
   margin-top: 13rem;
 }
 
-.srp-container {
+.srp-container, .srp-container-mobile {
   margin-top: 14rem;
 }
