+++ src/scss/productDetailsPage/productDetails.scss
@@ -115,3 +115,37 @@
         }
     }
 }
+
+// MOBILE & TABLET
+
+.pdp-mobile{
+    &__wrapper{
+        display: flex;
+        flex-direction: column;
+    }
+    &__backBtn{
+        top: 0.5rem;
+        right: 0.5rem;
+    }
+    &__left{
+        width: 100%;
+    }
+    &__right{
+        width: 100%;
+    }
+    .auc-Recommend-container{
+    .auc-Recommend-list{
+    display: block;
+    flex-wrap: nowrap;
+    display: -webkit-box;
+    overflow-x: scroll;
+    padding: 1rem 0rem;
+    list-style-type: none;
+
+    .auc-Recommend-item{
+        width: 40%;
+    }
+    }
+}
+
+}
