+++ src/scss/recommend/recommend.scss
@@ -15,7 +15,6 @@
         flex-wrap: nowrap;
         gap: 0.5rem;
         // display: -webkit-box;
-        overflow-x: scroll;
         padding: 1rem 0rem;
         list-style-type: none;
         align-items: center;
