+++ src/scss/searchresultspage/searchresultspage.scss
@@ -116,3 +116,57 @@
     }
   }
 }
+
+
+// ------------------------------------------
+// SRP Container Mobile & tablet
+// ------------------------------------------
+
+.srp-container-mobile{
+  width: 95%;
+  margin: 1rem auto;
+  display: flex;
+  gap: 3rem;
+  position: relative;
+
+  .srp-container__stats-sort{
+    margin: 1rem 0;
+  }
+
+  .srp-container__hits{
+    .ais-InfiniteHits{
+      .ais-InfiniteHits-list{
+        display: grid;
+        grid-template-columns: repeat(2, 1fr);
+      }
+    }
+  }
+
+  .facets-slider{
+    left: 1rem;
+    top: 3rem;
+    flex-direction: row;
+    align-items: center;
+    justify-content: center;
+    gap: 0.5rem;
+    display: flex;
+    position: absolute;
+    height: fit-content;
+    z-index: 999;
+    transition: 0.2s ease-in-out;
+
+    p{
+      font-size: 0.8rem;
+    }
+    svg{
+      // margin-top: 0.5rem;
+      width: 8px;
+      transition: transform 0.3s ease-in-out;
+    }
+  }
+  .facets-slider-active{
+  }
+  .facets-slider-inactive{
+
+  }
+}
